@@ -1,6 +1,6 @@
 {
   "name": "fargate-appmesh-cdk",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "bin": {
     "fargate-appmesh-cdk": "bin/fargate-appmesh-cdk.js"
   },
@@ -10,22 +10,22 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@types/node": "8.10.40",
-    "aws-cdk": "^0.28.0",
-    "typescript": "^3.3.3333"
+    "@types/node": "12.12.6",
+    "aws-cdk": "^1.23.0",
+    "typescript": "^3.6.0"
   },
   "dependencies": {
-    "@aws-cdk/aws-appmesh": "^0.28.0",
-    "@aws-cdk/aws-cloudwatch": "^0.28.0",
-    "@aws-cdk/aws-ec2": "^0.28.0",
-    "@aws-cdk/aws-ecr": "^0.28.0",
-    "@aws-cdk/aws-ecs": "^0.28.0",
-    "@aws-cdk/aws-elasticloadbalancingv2": "^0.28.0",
-    "@aws-cdk/aws-iam": "^0.28.0",
-    "@aws-cdk/aws-logs": "^0.28.0",
-    "@aws-cdk/aws-servicediscovery": "^0.28.0",
-    "@aws-cdk/aws-ssm": "^0.28.0",
-    "@aws-cdk/cdk": "^0.28.0",
-    "source-map-support": "^0.5.9"
+    "@aws-cdk/aws-appmesh": "^1.23.0",
+    "@aws-cdk/aws-cloudwatch": "^1.23.0",
+    "@aws-cdk/aws-ec2": "^1.23.0",
+    "@aws-cdk/aws-ecr": "^1.23.0",
+    "@aws-cdk/aws-ecs": "^1.23.0",
+    "@aws-cdk/aws-elasticloadbalancingv2": "^1.23.0",
+    "@aws-cdk/aws-iam": "^1.23.0",
+    "@aws-cdk/aws-logs": "^1.23.0",
+    "@aws-cdk/aws-servicediscovery": "^1.23.0",
+    "@aws-cdk/aws-ssm": "^1.23.0",
+    "@aws-cdk/core": "^1.23.0",
+    "source-map-support": "^0.5.16"
   }
 }
